Minor edits to functions - accounts for 2 timers
Added a 'display' argument to setFormatTimer(). This allows the function to be used in multiple locations at once. Also - modified each instance of setFormatTime() to account for said changes. SetFormatTime(duration, display)

Added rest_timer variable and rest_display variable. Not certain if the rest_timer variable is required or if we should just reset timer and use that one in multiple locations.